feat: implementa sistema completo de upload de fotos de perfil e notificações
- Adiciona serviço de upload de fotos de perfil com integração Supabase Storage
- Implementa sistema de notificações segmentadas com OneSignal
- Corrige permissões de storage e políticas RLS no Supabase
- Adiciona testes para widgets de trip e waiting driver screen
- Atualiza configurações de build para Android e iOS
- Implementa validação e tratamento de erros para upload de fotos
- Adiciona documentação técnica e guias de implementação
- Remove serviço FCM obsoleto e migra para OneSignal

+++ lib/services/cancellation_fee_service.dart
@@ -212,8 +212,8 @@ class CancellationFeeService {
         return 0.5; // 50% como fallback
       }
 
-      final factor = math.min(1, distanceTraveled / totalDistance);
-      return math.max(0.1, factor); // Mínimo 10%, máximo 100%
+      final factor = math.min(1.0, distanceTraveled / totalDistance);
+      return math.max(0.1, factor as double); // Mínimo 10%, máximo 100%
     } catch (e) {
       print('⚠️ Erro ao calcular fator de deslocamento: $e');
       return 0.5; // 50% como fallback seguro
